fix(): Created blueprint of /weather
